backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-settings>=2.6.0
pydantic>=2.9.0
supabase>=2.10.0
twilio>=9.3.0
google-generativeai>=0.8.0
python-multipart>=0.0.12
python-dotenv>=1.0.1
httpx>=0.27.0
celery>=5.4.0
redis>=5.0.0
